Add exception for runningInConsole in FilamentFabricatorServiceProvider.php
I've encountered an issue with the `Z3d0X/filament-fabricator` package in Laravel, related to using models directly within Block files. The problem arises when these models are loaded from a ServiceProvider. This becomes evident during a fresh migration process, where the error "table not exists" is thrown because the system attempts to access the table while it is being migrated.